+++ pymaltego/entities.py
@@ -1,7 +1,5 @@
 # coding=utf-8
 
-import json
-
 from lxml import etree
 
 from . import exceptions
@@ -11,14 +9,12 @@ class Node(object):
 
     """Node object."""
 
-    def __new__(self, name, value=None, parent=None, **kwargs):
-        """
-        Create object.
+    def __new__(cls, name, value=None, parent=None, **kwargs):
+        """Create object.
 
         :param name: name of node.
         :param value (optional): text of node.
         :param parent (optional): `etree.Element` instance parent of node.
-        :param kwargs (optional): `dict` of arguments.
         """
         element = etree.Element(name, **kwargs)
 
@@ -35,172 +31,161 @@ class XMLObject(object):
 
     """XML object."""
 
-    def __init__(self, node=None):
-        """
-        Initialization object.
-
-        :param node (optional): `etree.Element` instance.
-        """
-        if node is not None:
-            self.load_from_node(node)
-
-    def load_from_node(self, node):
-        """
-        Load values from node.
+    @classmethod
+    def from_node(cls, node):
+        """Load values from node.
 
         :param node: `etree.Element` instance.
         """
         if not etree.iselement(node):
             raise ValueError('Is not an `etree.Element` instance.')
 
+        if node.tag != cls.__name__:
+            raise exceptions.MalformedEntityError(
+                '{} not a "{}" tag.'.format(node.tag, cls.__name__)
+            )
+
     def to_node(self):
-        """
-        Serialize to `etree.Element` instance.
+        """Serialize to `etree.Element` instance.
 
         :returns: `etree.Element` instance.
         """
         raise NotImplementedError('Object should contains method `to_node`.')
 
     def to_dict(self):
-        """
-        Serialize to `dict` instance.
+        """Serialize to `dict` instance.
 
         :returns: `dict` instance.
         """
         raise NotImplementedError('Object should contains method `to_dict`.')
 
     def to_xml(self, pretty_print=False):
-        """
-        Serialize to XML string.
+        """Serialize to XML string.
 
         :returns: `str` XML.
         """
         return etree.tostring(self.to_node(), pretty_print=pretty_print)
-
-    def to_json(self, pretty_print=False):
-        """
-        Serialize to JSON string.
-
-        :returns: `str` JSON.
-        """
-        if pretty_print:
-            return json.dumps(self.to_dict(), sort_keys=True, indent=4)
-        return json.dumps(self.to_dict())
 
 
 class Label(XMLObject):
 
     """Label object."""
 
-    def __init__(self, name='', value='', content_type='text/html',
-                 node=None):
-        """
-        Initialization object.
+    def __init__(self, name, value, content_type='text/html'):
+        """Initialization object.
 
         :param name (optional): label name.
         :param value (optional): label value.
         :param content_type (optional): label content type.
-        :param node (optional): `etree.Element` instance.
         """
         self.name = name
         self.value = value
         self.content_type = content_type
-        return super(Label, self).__init__(node)
 
-    def load_from_node(self, node):
-        """
-        Load values from node.
+    @classmethod
+    def from_node(cls, node):
+        """Load values from node.
 
         :param node: `etree.Element` instance.
+        :returns: `entities.Label` instance.
         """
-        super(Label, self).load_from_node(node)
+        super(Label, cls).from_node(node)
 
-        if node.tag != self.__class__.__name__:
-            raise exceptions.MalformedEntityError(
-                '{} not a "{}" tag.'.format(node.tag, self.__class__.__name__)
-            )
-
-        self.name = node.attrib.get('Name', '')
-        self.content_type = node.attrib.get('Type', '')
-        self.value = node.text and node.text.strip()
+        return cls(
+            node.attrib['Name'], node.text.strip(), node.attrib.get('Type', '')
+        )
 
     def to_node(self):
-        """
-        Serialize to `etree.Element` instance.
+        """Serialize to `etree.Element` instance.
 
         :returns: `etree.Element` instance.
         """
-        node = Node('Label')
+        node = Node(self.__class__.__name__)
+
         node.attrib['Name'] = self.name
         node.attrib['Type'] = self.content_type
         node.text = etree.CDATA(self.value)
+
         return node
+
+    def to_dict(self):
+        """Serialize to `dict` instance.
+
+        :returns: `dict` instance.
+        """
+        return {
+            'name': self.name, 'value': self.value,
+            'content_type': self.content_type
+        }
 
 
 class Field(XMLObject):
 
     """Field object."""
 
-    def __init__(self, name='', value='', display_name='',
-                 matching_rule=None, node=None):
-        """
-        Initialization object.
+    def __init__(self, name, value, display_name=None, matching_rule=None):
+        """Initialization object.
 
         :param name (optional): field name.
         :param value (optional): field value.
         :param display_name (optional): field display name.
         :param matching_rule (optional): field matching rule.
-        :param node (optional): `etree.Element` instance.
         """
         self.name = name
         self.value = value
-        self.display_name = display_name
+        self.display_name = display_name or (
+            self.name[0].upper() + self.name[1:]
+        )
         self.matching_rule = matching_rule
-        super(Field, self).__init__(node)
 
-    def load_from_node(self, node):
-        """
-        Load values from node.
+    @classmethod
+    def from_node(cls, node):
+        """Load values from node.
 
         :param node: `etree.Element` instance.
+        :returns: `entities.Field` instance.
         """
-        super(Field, self).load_from_node(node)
+        super(Field, cls).from_node(node)
 
-        if node.tag != self.__class__.__name__:
-            raise exceptions.MalformedEntityError(
-                '{} not a "{}" tag'.format(node.tag, self.__class__.__name__)
-            )
-
-        self.name = node.attrib.get('Name')
-        self.value = node.text and node.text.strip()
-        self.display_name = node.attrib.get('DisplayName')
-        self.matching_rule = node.attrib.get('MatchingRule')
+        return cls(
+            node.attrib['Name'], node.text.strip(),
+            node.attrib.get('DisplayName'), node.attrib.get('MatchingRule')
+        )
 
     def to_node(self):
-        """
-        Serialize to `etree.Element` instance.
+        """Serialize to `etree.Element` instance.
 
         :returns: `etree.Element` instance.
         """
-        node = Node('Field', self.value)
-        node.attrib['Name'] = self.name
+        node = Node(self.__class__.__name__, self.value)
 
-        if self.display_name is not None:
-            node.attrib['DisplayName'] = self.display_name
+        node.attrib['Name'] = self.name
+        node.attrib['DisplayName'] = self.display_name
 
         if self.matching_rule is not None:
             node.attrib['MatchingRule'] = self.matching_rule
+
         return node
+
+    def to_dict(self):
+        """Serialize to `dict` instance.
+
+        :returns: `dict` instance.
+        """
+        return {
+            'name': self.name, 'value': self.value,
+            'display_name': self.display_name,
+            'matching_rule': self.matching_rule
+        }
 
 
 class Entity(XMLObject):
 
     """Entity base object."""
 
-    def __init__(self, name='', value='', weight=None, icon_url=None,
-                 labels=None, fields=None, node=None):
-        """
-        Initialization object.
+    def __init__(self, name, value, weight=None, icon_url=None,
+                 labels=None, fields=None):
+        """Initialization object.
 
         :param name (optional): entity name.
         :param value (optional): entity value.
@@ -208,68 +193,64 @@ class Entity(XMLObject):
         :param icon_url (optional): entity icon url.
         :param labels (optional): entity display information.
         :param fields (optional): `dict` of `Field` instance.
-        :param node (optional): `etree.Element` instance.
         """
         self.name = name
         self.value = value
         self.weight = weight
         self.icon_url = icon_url
-        self.fields = fields or {}
-        self.labels = labels or {}
+        self.fields = fields or []
+        self.labels = labels or []
 
-        self.node = node
-
-        super(Entity, self).__init__(node)
-
-    def load_from_node(self, node):
-        """
-        Load values from node.
+    @classmethod
+    def from_node(cls, node):
+        """Load values from node.
 
         :param node: `etree.Element` instance.
+        :returns: `entities.Entity` instance.
         """
-        super(Entity, self).load_from_node(node)
+        super(Entity, cls).from_node(node)
 
-        if node.tag != self.__class__.__name__:
-            raise exceptions.MalformedEntityError(
-                '{} not a {}.'.format(node.tag, self.__class__.__name__)
-            )
-
-        if 'Type' not in node.attrib:
+        try:
+            name = node.attrib['Type']
+        except KeyError:
             raise exceptions.MalformedEntityError('No "Type" attribute.')
-        self.name = node.attrib['Type']
 
         value = node.find('Value')
         if value is None:
             raise exceptions.MalformedEntityError('Missing "Value" tag.')
-        self.value = value.text and value.text.strip()
+
+        value = value.text and value.text.strip()
+
+        instance = cls(name, value)
 
         weight = node.find('Weight')
         if weight is not None:
-            self.weight = weight.text and weight.text.strip()
+            instance.weight = weight.text and weight.text.strip()
 
         additional_fields = node.find('AdditionalFields')
         if additional_fields is not None:
             for field_node in additional_fields.getchildren():
-                field = Field(node=field_node)
-                self.fields[field.name] = field
+                field = Field.from_node(field_node)
+                instance.fields.append(field)
 
         labels = node.find('DisplayInformation')
         if labels is not None:
             for label_node in labels.getchildren():
-                label = Label(node=label_node)
-                self.labels[label.name] = label
+                label = Label.from_node(label_node)
+                instance.labels.append(label)
 
         icon_url = node.find('IconURL')
         if icon_url is not None:
-            self.icon_url = icon_url.text and icon_url.text.strip()
+            instance.icon_url = icon_url.text and icon_url.text.strip()
+
+        return instance
 
     def to_node(self):
-        """
-        Serialize to `etree.Element` instance.
+        """Serialize to `etree.Element` instance.
 
         :returns: `etree.Element` instance.
         """
-        node = Node('Entity')
+        node = Node(self.__class__.__name__)
         node.attrib['Type'] = self.name
 
         Node('Value', self.value, parent=node)
@@ -279,23 +260,22 @@ class Entity(XMLObject):
 
         if self.fields:
             additional_fields = Node('AdditionalFields', parent=node)
-            for field in self.fields.values():
+            for field in self.fields:
                 if field.value:
                     additional_fields.append(field.to_node())
 
         if self.labels:
             labels = Node('DisplayInformation', parent=node)
-            for label in self.labels.values():
+            for label in self.labels:
                 labels.append(label.to_node())
 
-        if self.icon_url is not None:
+        if self.icon_url:
             Node('IconURL', value=self.icon_url, parent=node)
 
         return node
 
     def to_dict(self):
-        """
-        Serialize to `dict` instance.
+        """Serialize to `dict` instance.
 
         :returns: `dict` instance.
         """
@@ -305,7 +285,7 @@ class Entity(XMLObject):
         result[self.name]['weight'] = self.weight
 
         result[self.name]['fields'] = {
-            key: field.value for key, field in self.fields.items()
+            field.name.lower(): field.value for field in self.fields
         }
 
         return result