+++ tests.py
@@ -1,7 +1,6 @@
 # coding=utf-8
 
 import unittest
-import json
 
 from lxml import etree
 
@@ -51,21 +50,14 @@ class XMLObjectTests(unittest.TestCase):
 
         self.assertIsInstance(xml_object, entities.XMLObject)
 
-    def test_create__with_node(self):
-        """Testing create instance with node."""
-        node = entities.Node('Test')
-        xml_object = entities.XMLObject(node=node)
-
-        self.assertIsInstance(xml_object, entities.XMLObject)
-
     def test_create__with_wrong_node(self):
         """Testing create instance with wrong node."""
         node = 'Test'
 
         with self.assertRaises(ValueError):
-            entities.XMLObject(node=node)
+            entities.XMLObject.from_node(node)
 
-    def test_method_to_node_not_implemented(self):
+    def test_method_to_node(self):
         """Testing method raises NotImplementedError."""
         xml_object = entities.XMLObject()
 
@@ -93,35 +85,26 @@ class LabelTests(unittest.TestCase):
 
     def test_create(self):
         """Testing create instance."""
-        label = entities.Label()
+        name = value = 'Test'
+        label = entities.Label(name=name, value=value)
 
         self.assertIsInstance(label, entities.Label)
-
-    def test_create__with_name(self):
-        """Testing create instance with name."""
-        name = 'Test'
-        label = entities.Label(name=name)
-
         self.assertEqual(label.name, name)
-
-    def test_create__with_value(self):
-        """Testing create instance with value."""
-        value = 'Test'
-        label = entities.Label(value=value)
-
         self.assertEqual(label.value, value)
 
     def test_create__with_content_type(self):
         """Testing create instance with content type."""
-        content_type = 'Test'
-        label = entities.Label(content_type=content_type)
+        name = value = content_type = 'Test'
+        label = entities.Label(
+            name=name, value=value, content_type=content_type
+        )
 
         self.assertEqual(label.content_type, content_type)
 
-    def test_create__with_node(self):
+    def test_create_from_node(self):
         """Testing create instance with node."""
-        node = entities.Node('Label')
-        label = entities.Label(node=node)
+        node = entities.Node('Label', 'Test', attrib={'Name': 'Test'})
+        label = entities.Label.from_node(node=node)
 
         self.assertIsInstance(label, entities.Label)
 
@@ -130,7 +113,7 @@ class LabelTests(unittest.TestCase):
         node = entities.Node('Wrong')
 
         with self.assertRaises(exceptions.MalformedEntityError):
-            entities.Label(node=node)
+            entities.Label.from_node(node=node)
 
     def test_method_to_node(self):
         """Testing method to_node."""
@@ -158,49 +141,53 @@ class LabelTests(unittest.TestCase):
             ).format(name)
         )
 
+    def test_method_to_dict(self):
+        """Testing method to_dict."""
+        name = value = content_type = 'Test'
+        label_dict = entities.Label(
+            name=name, value=value, content_type=content_type
+        ).to_dict()
+
+        self.assertEqual(name, label_dict['name'])
+        self.assertEqual(value, label_dict['value'])
+        self.assertEqual(value, label_dict['value'])
+        self.assertEqual(content_type, label_dict['content_type'])
+
 
 class FieldTests(unittest.TestCase):
 
     """Testing `entities.Field` object."""
 
     def test_create(self):
-        """Testing create instance."""
-        field = entities.Field()
-
-        self.assertIsInstance(field, entities.Field)
-
-    def test_create__with_name(self):
         """Testing create instance with name."""
-        name = 'Test'
-        field = entities.Field(name=name)
+        name = value = 'Test'
+        field = entities.Field(name=name, value=value)
 
         self.assertEqual(field.name, name)
-
-    def test_create__with_value(self):
-        """Testing create instance with value."""
-        value = 'Test'
-        field = entities.Field(value=value)
-
-        self.assertEqual(field.value, value)
+        self.assertEqual(field.display_name, name)
 
     def test_create__with_display_name(self):
         """Testing create instance with display name."""
-        display_name = 'Test'
-        field = entities.Field(display_name=display_name)
+        name = value = display_name = 'Test'
+        field = entities.Field(
+            name=name, value=value, display_name=display_name
+        )
 
         self.assertEqual(field.display_name, display_name)
 
     def test_create__with_matching_rule(self):
         """Testing create instance with matching_rule."""
-        matching_rule = 'Test'
-        field = entities.Field(matching_rule=matching_rule)
+        name = value = matching_rule = 'Test'
+        field = entities.Field(
+            name=name, value=value, matching_rule=matching_rule
+        )
 
         self.assertEqual(field.matching_rule, matching_rule)
 
     def test_create__with_node(self):
         """Testing create instance with node."""
-        node = entities.Node('Field')
-        field = entities.Field(node=node)
+        node = entities.Node('Field', 'Test', attrib={'Name': 'Test'})
+        field = entities.Field.from_node(node=node)
 
         self.assertIsInstance(field, entities.Field)
 
@@ -209,7 +196,7 @@ class FieldTests(unittest.TestCase):
         node = entities.Node('Wrong')
 
         with self.assertRaises(exceptions.MalformedEntityError):
-            entities.Field(node=node)
+            entities.Field.from_node(node=node)
 
     def test_method_to_node(self):
         """Testing method to_node."""
@@ -240,6 +227,20 @@ class FieldTests(unittest.TestCase):
             ).format(name)
         )
 
+    def test_method_to_dict(self):
+        """Testing method to_dict."""
+        name = value = display_name = matching_rule = 'Test'
+        field_dict = entities.Field(
+            name=name, value=value, display_name=display_name,
+            matching_rule=matching_rule
+        ).to_dict()
+
+        self.assertEqual(name, field_dict['name'])
+        self.assertEqual(value, field_dict['value'])
+        self.assertEqual(display_name, field_dict['display_name'])
+        self.assertEqual(display_name, field_dict['display_name'])
+        self.assertEqual(matching_rule, field_dict['matching_rule'])
+
 
 class EntityTests(unittest.TestCase):
 
@@ -247,49 +248,40 @@ class EntityTests(unittest.TestCase):
 
     def test_create(self):
         """Testing create instance."""
-        entity = entities.Entity()
+        name = value = 'Test'
+        entity = entities.Entity(name=name, value=value)
 
         self.assertIsInstance(entity, entities.Entity)
-
-    def test_create__with_name(self):
-        """Testing create instance with name."""
-        name = 'Test'
-        entity = entities.Entity(name=name)
-
         self.assertEqual(entity.name, name)
-
-    def test_create__with_value(self):
-        """Testing create instance with value."""
-        value = 'Test'
-        entity = entities.Entity(value=value)
-
         self.assertEqual(entity.value, value)
 
     def test_create__with_weight(self):
         """Testing create instance with weight."""
-        weight = 'Test'
-        entity = entities.Entity(weight=weight)
+        name = value = weight = 'Test'
+        entity = entities.Entity(name=name, value=value, weight=weight)
 
         self.assertEqual(entity.weight, weight)
 
     def test_create__with_icon_url(self):
         """Testing create instance with icon_url."""
-        icon_url = 'Test'
-        entity = entities.Entity(icon_url=icon_url)
+        name = value = icon_url = 'Test'
+        entity = entities.Entity(name=name, value=value, icon_url=icon_url)
 
         self.assertEqual(entity.icon_url, icon_url)
 
     def test_create__with_labels(self):
         """Testing create instance with labels."""
-        labels = {'Test': 'Test'}
-        entity = entities.Entity(labels=labels)
+        name = value = 'Test'
+        labels = []
+        entity = entities.Entity(name=name, value=value, labels=labels)
 
         self.assertEqual(entity.labels, labels)
 
     def test_create__with_fields(self):
         """Testing create instance with fields."""
-        fields = {'Test': 'Test'}
-        entity = entities.Entity(fields=fields)
+        name = value = 'Test'
+        fields = []
+        entity = entities.Entity(name=name, value=value, fields=fields)
 
         self.assertEqual(entity.fields, fields)
 
@@ -297,7 +289,7 @@ class EntityTests(unittest.TestCase):
         """Testing create instance with node."""
         node = entities.Node('Entity', attrib={'Type': 'Test'})
         entities.Node('Value', parent=node)
-        entity = entities.Entity(node=node)
+        entity = entities.Entity.from_node(node=node)
 
         self.assertIsInstance(entity, entities.Entity)
 
@@ -306,55 +298,55 @@ class EntityTests(unittest.TestCase):
         node = entities.Node('Wrong')
 
         with self.assertRaises(exceptions.MalformedEntityError):
-            entities.Entity(node=node)
+            entities.Entity.from_node(node=node)
 
-    def test_create__with_wrong_node_without_type(self):
+    def test_create__wrong_node_type(self):
         """Testing create instance with wrong node without type."""
         node = entities.Node('Entity')
 
         with self.assertRaises(exceptions.MalformedEntityError):
-            entities.Entity(node=node)
+            entities.Entity.from_node(node=node)
 
-    def test_create__with_wrong_node_without_value(self):
+    def test_create__wrong_node_value(self):
         """Testing create instance with wrong node without value."""
         node = entities.Node('Entity', attrib={'Type': 'Test'})
 
         with self.assertRaises(exceptions.MalformedEntityError):
-            entities.Entity(node=node)
+            entities.Entity.from_node(node=node)
 
-    def test_create__with_node_weight(self):
+    def test_create__node_weight(self):
         """Testing create instance with node weight."""
         weight = '0'
         node = entities.Node('Entity', attrib={'Type': 'Test'})
         entities.Node('Value', parent=node)
         entities.Node('Weight', value=weight, parent=node)
-        entity = entities.Entity(node=node)
+        entity = entities.Entity.from_node(node=node)
 
         self.assertEqual(entity.weight, weight)
 
-    def test_create__with_node_additional_fields(self):
+    def test_create__additional_fields(self):
         """Testing create instance with node additional fields."""
         field_name = 'Test'
         node = entities.Node('Entity', attrib={'Type': 'Test'})
         entities.Node('Value', parent=node)
 
         fields = entities.Node('AdditionalFields', parent=node)
-        entities.Node('Field', attrib={'Name': field_name}, parent=fields)
-        entity = entities.Entity(node=node)
+        entities.Node('Field', 'V', attrib={'Name': field_name}, parent=fields)
+        entity = entities.Entity.from_node(node=node)
 
-        self.assertIn(field_name, entity.fields)
+        self.assertEqual(len(entity.fields), 1)
 
     def test_create__with_node_labels(self):
         """Testing create instance with node labels."""
         label_name = 'Test'
-        node = entities.Node('Entity', attrib={'Type': 'Test'})
+        node = entities.Node('Entity', 'Test', attrib={'Type': 'Test'})
         entities.Node('Value', parent=node)
 
         labels = entities.Node('DisplayInformation', parent=node)
-        entities.Node('Label', attrib={'Name': label_name}, parent=labels)
-        entity = entities.Entity(node=node)
+        entities.Node('Label', 'V', attrib={'Name': label_name}, parent=labels)
+        entity = entities.Entity.from_node(node=node)
 
-        self.assertIn(label_name, entity.labels)
+        self.assertEqual(len(entity.labels), 1)
 
     def test_create__with_node_icon_url(self):
         """Testing create instance with node icon url."""
@@ -363,18 +355,18 @@ class EntityTests(unittest.TestCase):
         entities.Node('Value', parent=node)
 
         entities.Node('IconURL', value=value, parent=node)
-        entity = entities.Entity(node=node)
+        entity = entities.Entity.from_node(node=node)
 
         self.assertEqual(entity.icon_url, value)
 
     def test_method_to_node(self):
         """Testing method to_node."""
         name = value = weight = icon_url = 'Test'
-        field = entities.Field(value=value)
-        label = entities.Label(value=value)
+        field = entities.Field(name=name, value=value)
+        label = entities.Label(name=name, value=value)
         node = entities.Entity(
             name=name, value=value, weight=weight,
-            icon_url=icon_url, fields={'Field': field}, labels={'Label': label}
+            icon_url=icon_url, fields=[field], labels=[label]
         ).to_node()
 
         self.assertEqual(node.attrib['Type'], name)
@@ -391,51 +383,17 @@ class EntityTests(unittest.TestCase):
     def test_method_to_dict(self):
         """Testing method to_dict."""
         name = value = weight = icon_url = 'Test'
-        field = entities.Field(value=value)
-        label = entities.Label(value=value)
-        d = entities.Entity(
+        field = entities.Field(name=name, value=value)
+        label = entities.Label(name=name, value=value)
+        entity_dict = entities.Entity(
             name=name, value=value, weight=weight,
-            icon_url=icon_url, fields={'Field': field}, labels={'Label': label}
+            icon_url=icon_url, fields=[field], labels=[label]
         ).to_dict()
 
-        self.assertIn(name, d)
-        self.assertIn('fields', d[name])
-        self.assertIn('Field', d[name]['fields'])
-        self.assertEqual(d[name]['fields']['Field'], value)
-
-    def test_method_to_json(self):
-        """Testing method to_dict."""
-        name = value = weight = icon_url = 'Test'
-        field = entities.Field(value=value)
-        label = entities.Label(value=value)
-        json_string = entities.Entity(
-            name=name, value=value, weight=weight,
-            icon_url=icon_url, fields={'Field': field}, labels={'Label': label}
-        ).to_json()
-
-        d = json.loads(json_string)
-
-        self.assertIn(name, d)
-        self.assertIn('fields', d[name])
-        self.assertIn('Field', d[name]['fields'])
-        self.assertEqual(d[name]['fields']['Field'], value)
-
-    def test_method_to_json__pretty_print(self):
-        """Testing method to_dict."""
-        name = value = weight = icon_url = 'Test'
-        field = entities.Field(value=value)
-        label = entities.Label(value=value)
-        json_string = entities.Entity(
-            name=name, value=value, weight=weight,
-            icon_url=icon_url, fields={'Field': field}, labels={'Label': label}
-        ).to_json(pretty_print=True)
-
-        d = json.loads(json_string)
-
-        self.assertIn(name, d)
-        self.assertIn('fields', d[name])
-        self.assertIn('Field', d[name]['fields'])
-        self.assertEqual(d[name]['fields']['Field'], value)
+        self.assertIn(name, entity_dict)
+        self.assertIn('fields', entity_dict[name])
+        self.assertIn(name.lower(), entity_dict[name]['fields'])
+        self.assertEqual(entity_dict[name]['fields'][name.lower()], value)
 
 
 class TransformRequestTests(unittest.TestCase):
@@ -452,7 +410,7 @@ class TransformRequestTests(unittest.TestCase):
         )
         entities.Node('Value', value='Test', parent=entity)
 
-        message = messages.TransformRequest(node=node)
+        message = messages.TransformRequest.from_node(node=node)
 
         self.assertIsInstance(message, messages.TransformRequest)
 
@@ -468,7 +426,7 @@ class TransformRequestTests(unittest.TestCase):
         fields = entities.Node('TransformFields', parent=request)
         entities.Node('Field', attrib={'Name': 'Test'}, parent=fields)
 
-        message = messages.TransformRequest(node=node)
+        message = messages.TransformRequest.from_node(node=node)
 
         self.assertIsInstance(message, messages.TransformRequest)
 
@@ -483,27 +441,34 @@ class TransformRequestTests(unittest.TestCase):
         entities.Node('Value', value='Test', parent=entity)
         entities.Node('Limits', parent=request)
 
-        message = messages.TransformRequest(node=node)
+        message = messages.TransformRequest.from_node(node=node)
 
         self.assertIsInstance(message, messages.TransformRequest)
 
     def test_create__with_wrong_node(self):
         """Testing create instance with wrong node."""
+        node = 'Test'
+
+        with self.assertRaises(ValueError):
+            messages.TransformRequest.from_node(node)
+
+    def test_create__with_wrong_tag(self):
+        """Testing create instance with wrong node."""
         node = entities.Node('MaltegoMessage')
         entities.Node('MaltegoWrongMessage', parent=node)
 
         with self.assertRaises(exceptions.MalformedMessageError):
-            messages.TransformRequest(node=node)
+            messages.TransformRequest.from_node(node=node)
 
-    def test_create__with_wrong_node_without_entities(self):
+    def test_create__without_entities(self):
         """Testing create instance with wrong node without entities."""
         node = entities.Node('MaltegoMessage')
         entities.Node('MaltegoTransformRequestMessage', parent=node)
 
         with self.assertRaises(exceptions.MalformedMessageError):
-            messages.TransformRequest(node=node)
+            messages.TransformRequest.from_node(node=node)
 
-    def test_create__with_wrong_node_without_field_name(self):
+    def test_create__without_field_name(self):
         """Testing create instance with wrong node without field name."""
         node = entities.Node('MaltegoMessage')
         request = entities.Node('MaltegoTransformRequestMessage', parent=node)
@@ -516,31 +481,7 @@ class TransformRequestTests(unittest.TestCase):
         entities.Node('Field', parent=fields)
 
         with self.assertRaises(exceptions.MalformedMessageError):
-            messages.TransformRequest(node=node)
-
-
-class JSONTransformRequestTests(unittest.TestCase):
-
-    """Testing `messages.JSONTransformRequest`."""
-
-    def test_create__single_values(self):
-        """Testing create instance."""
-        data = 'value'
-
-        message = messages.JSONTransformRequest(data=data)
-
-        self.assertIsInstance(message, messages.JSONTransformRequest)
-        self.assertEqual(len(message.entities), 1)
-        self.assertEqual(message.entities[0].value, data)
-
-    def test_create__few_values(self):
-        """Testing create instance."""
-        data = '["test","value"]'
-
-        message = messages.JSONTransformRequest(data=data)
-
-        self.assertIsInstance(message, messages.JSONTransformRequest)
-        self.assertEqual(len(message.entities), 2)
+            messages.TransformRequest.from_node(node=node)
 
 
 class TransformResponseTests(unittest.TestCase):
@@ -549,10 +490,10 @@ class TransformResponseTests(unittest.TestCase):
 
     def test_create(self):
         """Testing create instance with entities."""
-        entities = ['Entity']
-        response = messages.TransformResponse(entities)
+        entities_iter = ['Entity']
+        response = messages.TransformResponse(entities_iter)
 
-        self.assertEqual(response.entities, entities)
+        self.assertEqual(response.entities, entities_iter)
 
     def test_to_node(self):
         """Testing `to_node` method."""
@@ -568,14 +509,14 @@ class TransformResponseTests(unittest.TestCase):
     def test_to_node_with_ui_messages(self):
         """Testing `to_node` method with ui messages."""
         entity_value = 'Test'
-        ui_messages = {'Key': ['Value']}
+        ui_messages = [{'type': 'Type', 'value': 'Value'}]
         entity = entities.Entity(name='Test', value=entity_value, labels=None)
         node = messages.TransformResponse(
             [entity], ui_messages
         ).to_node()
 
         self.assertEqual(
-            ui_messages['Key'][0],
+            ui_messages[0]['value'],
             node.find('UIMessages').getchildren()[0].text
         )
 
@@ -584,10 +525,10 @@ class TransformResponseTests(unittest.TestCase):
         entity = entities.Entity(
             name='Test', value='Test'
         )
-        d = messages.TransformResponse([entity]).to_dict()
+        response_dict = messages.TransformResponse([entity]).to_dict()
 
-        self.assertIn('entities', d)
-        self.assertEqual(d['entities'], [entity.to_dict()])
+        self.assertIn('entities', response_dict)
+        self.assertEqual(response_dict['entities'], [entity.to_dict()])
 
     def test_to_xml(self):
         """Testing to_xml method."""
@@ -620,15 +561,7 @@ class BaseTransformTests(unittest.TestCase):
         )
         entities.Node('Value', value='Test', parent=entity)
 
-        message = messages.TransformRequest(node=node)
-        transform = transforms.BaseTransform(message)
-
-        self.assertIsInstance(transform, transforms.BaseTransform)
-
-    def test_create__JSONTransfromRequest(self):
-        """Testing create instance with JSONTransformRequest."""
-
-        message = messages.JSONTransformRequest(data='value')
+        message = messages.TransformRequest.from_node(node=node)
         transform = transforms.BaseTransform(message)
 
         self.assertIsInstance(transform, transforms.BaseTransform)
@@ -639,7 +572,7 @@ class BaseTransformTests(unittest.TestCase):
         with self.assertRaises(ValueError):
             transforms.BaseTransform('wrong')
 
-    def test_method_transform__not_implemented(self):
+    def test_method_transform(self):
         """Testing call not implemented method transform."""
         node = entities.Node('MaltegoMessage')
         request = entities.Node('MaltegoTransformRequestMessage', parent=node)
@@ -649,7 +582,7 @@ class BaseTransformTests(unittest.TestCase):
         )
         entities.Node('Value', value='Test', parent=entity)
 
-        message = messages.TransformRequest(node=node)
+        message = messages.TransformRequest.from_node(node=node)
         transform = transforms.BaseTransform(message)
 
         with self.assertRaises(NotImplementedError):
@@ -665,7 +598,7 @@ class BaseTransformTests(unittest.TestCase):
         )
         entities.Node('Value', value='Test', parent=entity)
 
-        message = messages.TransformRequest(node=node)
+        message = messages.TransformRequest.from_node(node=node)
         transform = transforms.BaseTransform(message)
 
         with self.assertRaises(NotImplementedError):